feat(ble)!: include a service in a server service and widen the attribute count
Completes included service support, which until now existed only on the client
side: a server could not declare that one of its services includes another.

GattServerService could not simply become a node of a second list. It is already
an infra::IntrusiveForwardList<GattServerService>::NodeType for the server's own
list of services, and one object cannot be a node of two lists of the same type.
GattServerIncludedService is therefore its own node, referring to the included
service rather than owning it, which also keeps an include from implying a
lifetime relationship the specification does not have.

GetAttributeCount widens from uint8_t to uint16_t on GattService,
GattServerService and GattServerCharacteristic. The old type capped a service at
255 attributes where ATT handles run to 0xFFFF, and every include pushes a service
one closer to that cap. A test builds 128 characteristics, which is 257
attributes: under the old type that wrapped to 1.

The two changes land together because both touch the same accounting, and
rewriting it twice is worse than once.

BREAKING CHANGE: GetAttributeCount returns uint16_t rather than uint8_t on
GattService, GattServerService and GattServerCharacteristic.

+++ services/ble/GattTypes.cpp
@@ -193,7 +193,7 @@ namespace services
         return endHandle;
     }
 
-    uint8_t GattService::GetAttributeCount() const
+    uint16_t GattService::GetAttributeCount() const
     {
         return 0;
     }
